ButterflyEngineering MERGE wth master
   This is a massive merge with the butterflyEngineering branch
   -- It is not 100% complete but produces a model that seems
      to be valid for the default case.
   -- expect further work to allow Box/Pancake/other moderators
   -- pbip is not currently working [update later]

+++ Model/moderator/VacVessel.cxx
@@ -418,7 +418,8 @@ VacVessel::createLinks()
 
   // set Links:
   for(size_t i=2;i<6;i++)
-    FixedComp::setConnect(i,getSurfacePoint(4,i+1),getDirection(i));
+    FixedComp::setConnect(i,getSurfacePoint(4,static_cast<long int>(i+1)),
+			  getDirection(i));
 
   // Set Connect surfaces:
   for(int i=2;i<6;i++)
